% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{all_fields}
\alias{all_fields}
\title{All Field Names in a clinicaltrials.gov Record}
\format{
A character vector with the names of all 322 queriable fields
}
\source{
\url{https://clinicaltrials.gov/api/info/study_fields_list?fmt=JSON}
}
\usage{
data(all_fields)
}
\description{
Character vector containing all queriable field names in a clinicaltrials.gov
record
}
\examples{
all_fields
}
\keyword{fields}
