% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fields_vector-data.R
\docType{data}
\name{all_fields_vector}
\alias{all_fields_vector}
\title{All Field Names in a clinicaltrials.gov Record}
\format{
A character vector with the names of all queriable fields
}
\source{
\url {https://clinicaltrials.gov/api/info/study_fields_list?fmt=JSON}
}
\usage{
data(all_fields_vector)
}
\description{
All queriable field names in a clinicaltrials.gov record
}
\examples{
all_fields_vector
}
\keyword{fields}
