% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{baseline_characteristics_fields}
\alias{baseline_characteristics_fields}
\title{Baseline Measures and Characteristics Fields}
\format{
A character vector
}
\source{
\url{https://prsinfo.clinicaltrials.gov/results_definitions.html#Result_Baseline
}
\usage{
data(baseline_characteristics_fields)
}
\description{
Character vector containing a small selection of fields providing information regarding
the baseline measures and data collected in the study
}
\examples{
baseline_characteristics_fields
}
\keyword{fields}
