% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors_help-data.R
\docType{data}
\name{core_info_fields}
\alias{core_info_fields}
\title{Core Study Information Fields}
\format{
A character vector with a curated selection of 22 fields
}
\source{
\url
}
\usage{
data(core_info_fields)
}
\description{
Character vector containing a small selection of relevant fields which prvide
an overview of the studies
}
\examples{
core_info_fields
}
\keyword{fields}
