% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_handler.R
\name{csv_handler}
\alias{csv_handler}
\title{Requests clinicaltrials.gov API and Returns Parsed csv Response}
\usage{
csv_handler(url)
}
\arguments{
\item{url}{An (almost, unclear how strict this is) ASCII formatted string used
to query the clinicaltrials.gov API}
}
\description{
This function handles parsing of csv format API response.Returning a read csv.
Uses httr, readr and stringi.
}
\examples{
csv_handler(url)

}
\keyword{csv}
\keyword{format}
\keyword{response}
