% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{eligibility_fields}
\alias{eligibility_fields}
\title{Eligibility and Population Fields}
\format{
A character vector
}
\source{
\url{https://prsinfo.clinicaltrials.gov/definitions.html#Eligibility
}
\usage{
data(eligibility_fields)
}
\description{
Character vector containing a small selection of fields providing information regarding
eligibility requirements
}
\examples{
eligibility_fields
}
\keyword{fields}
