% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{extended_info_fields}
\alias{extended_info_fields}
\title{Extended Study Information Fields}
\format{
A character vector
}
\usage{
data(extended_info_fields)
}
\description{
Character vector containing a large selection of fields aimed at providing
a complete vision of the studies
}
\examples{
extended_info_fields
}
\keyword{fields}
