% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_lists_df-help-data.R
\docType{data}
\name{field_lists_df}
\alias{field_lists_df}
\title{Pre Loaded Field Lists, Description and Help}
\format{
Dataframe
}
\usage{
data(field_lists_df)
}
\description{
A dataframe with the name of available field lists, a short description,
their length, the fields included, and a help link.
}
\examples{
field_lists_df
}
\keyword{description}
\keyword{lists}
