% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{for_explorer}
\alias{for_explorer}
\title{Data fields input for the data explorer shiny app held in "AdlCruz/rctexplorer"}
\format{
A character vector
}
\usage{
data(for_explorer)
}
\description{
Character vector containing fields used in the data explorer app.
}
\examples{
for_explorer
}
\keyword{fields}
