% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get Up To 20 Study Fields}
\usage{
get_fields(
  search_expr,
  fields,
  max_studies = 30,
  format = "csv",
  just_fields = TRUE
)
}
\arguments{
\item{search_expr}{A string following the querying guidelines on clinicaltrials.gov}

\item{fields}{A list or character vector with up to 20 field names}

\item{max_studies}{A number indicating how many studies of all the ones that
match the search expression will be returned}
}
\description{
This function takes a search expression and character vector of 20 or less
field names and returns one dataframe with the parsed API response. Uses glue
}
\examples{
get_fields(expr = 'heart attack', fields = basic_info_fields())
}
\keyword{A}
