% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_content.R
\name{get_response_content}
\alias{get_response_content}
\title{Get Content of API Response}
\usage{
get_response_content(
  search_expr = search_expr,
  fields = fields,
  max_studies = max_studies
)
}
\arguments{
\item{search_expr}{A string following the querying guidelines at
https://clinicaltrials.gov/api/gui/ref/syntax and https://clinicaltrials.gov/api/gui/ref/expr}

\item{fields}{A character vector with field names. All field names available at
https://clinicaltrials.gov/api/info/study_fields_list?fmt=JSON, with "all_fields",
or with "get_vector_all_study_fields()"}

\item{max_studies}{A number indicating how many studies of all the ones that
match the search expression will be returned.}
}
\description{
Helper function for response_content parameter == TRUE. Returns decoded content of query respone.
Limited to 20 fields.
}
\keyword{20}
\keyword{fields}
\keyword{get}
\keyword{json}
\keyword{max}
\keyword{raw}
