% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_fields.R
\name{get_study_fields}
\alias{get_study_fields}
\title{Get Over 20 Study Fields}
\usage{
get_study_fields(search_expr, fields, max_studies = 30)
}
\arguments{
\item{search_expr}{A string following the querying guidelines on clinicaltrials.gov.}

\item{fields}{A character vector with any field names found on clinicaltrials.gov.}

\item{max_studies}{A number indicating how many studies of all the ones that
match the search expression will be returned.}
}
\description{
This function takes a search expression and character vector with any number
of field names and returns one dataframe with the merged and parsed csv API response.
}
\examples{
get_study_fields(expr = 'heart attack',fields = basic_info_and_results_fields())
}
\keyword{fields}
\keyword{get}
\keyword{many}
