% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vector_all_study_fields.R
\name{get_vector_all_study_fields}
\alias{get_vector_all_study_fields}
\title{Get All Study Fields Names}
\usage{
get_vector_all_study_fields()
}
\description{
This function returns a character vector with all 322 (May 2020) queriable field.
Uses glue.
names
}
\examples{
get_vector_all_study_fields()
}
\keyword{argument}
\keyword{fields}
