% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors_help-data.R
\docType{data}
\name{identification_and_status_fields}
\alias{identification_and_status_fields}
\title{Study Identification and Status Fields}
\format{
A character vector with a curated selection of 21 fields
}
\source{
\url https://prsinfo.clinicaltrials.gov/definitions.html#identification
}
\usage{
data(identification_and_status)
}
\description{
Character vector containing a small selection of fields aimed at allowing
identification of studies and corresponding status
}
\examples{
identification_and_status_fields
}
\keyword{fields}
