% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handler.R
\name{json_handler}
\alias{json_handler}
\title{Requests clinicaltrials.gov API and Returns Parsed json Response}
\usage{
json_handler(url)
}
\arguments{
\item{url}{An (almost, unclear how strict this is) ASCII formatted string used
to query the clinicaltrials.gov API}
}
\description{
This function handles parsing of json format API response.Returning a read csv.
Uses jsonlite.
}
\examples{
json_handler(url)

}
\keyword{format}
\keyword{json}
\keyword{response}
