% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{outcome_measures_info_fields}
\alias{outcome_measures_info_fields}
\title{Outcome Measures Description Fields}
\format{
A character vector with a curated selection of 9 fields
}
\source{
\url{https://prsinfo.clinicaltrials.gov/definitions.html#Outcomes
}
\usage{
data(outcome_measures_info_fields)
}
\description{
Character vector containing a small selection of fields providing information about
all outcome measures used
}
\examples{
outcome_measures_info_fields
}
\keyword{fields}
