% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors_help-data.R
\docType{data}
\name{outcome_measures_results_fields}
\alias{outcome_measures_results_fields}
\title{Outcome Measures and Statistical Analysis Fields}
\format{
A character vector with a selection of 38 fields
}
\source{
\url https://prsinfo.clinicaltrials.gov/results_definitions.html#Result_Outcome_Measure
}
\usage{
data(outcome_measures_results_fields)
}
\description{
Character vector containing a selection of fields providing information regarding
the outcomes measured and statistical analyses performed as well as the resulting values.
}
\examples{
outcome_measures_results_fields
}
\keyword{fields}
