% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{participant_flow_fields}
\alias{participant_flow_fields}
\title{Participant Flow Fields}
\format{
A character vector with a selection of 14 fields
}
\source{
\url{https://prsinfo.clinicaltrials.gov/results_definitions.html#Result_ParticipantFlow
}
\usage{
data(participant_flow_fields)
}
\description{
Character vector containing a small selection of fields providing information regarding
the progress of participants through each stage of a study
}
\examples{
participant_flow_fields
}
\keyword{fields}
