% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_ct.R
\name{request_ct}
\alias{request_ct}
\title{Requests clinicaltrials.gov API and Returns Response}
\usage{
request_ct(url)
}
\arguments{
\item{url}{A formatted string used to query the clinicaltrials.gov API}
}
\description{
This function requests the official clinicaltrials.gov API and returns an unparsed
response.
}
\examples{
request_ct(url)
}
\keyword{request}
\keyword{response}
\keyword{url}
