% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_vector_unique.R
\name{split_vector_unique}
\alias{split_vector_unique}
\title{Splits Fields Character Vector}
\usage{
split_vector_unique(x, n)
}
\arguments{
\item{x}{A character vector}

\item{n}{The number of elements (strings) in each column}
}
\description{
This function splits a character vector of any length into a matrix with n rows
of x elements. This is to get around the 20 field limit imposed by the API.
When necessary removes duplicates and fills columns with NA values.
Adapted from https://stackoverflow.com/a/48930537
}
\examples{
split_vector_unique(x= sample(c("April", "is", "the","cruellest","month","breeding",
"lilacs"), size = 40, replace = TRUE), n =5)

}
\keyword{split}
\keyword{unique}
\keyword{vector}
