% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors_help-data.R
\docType{data}
\name{study_design_arms_groups_and_interventions_fields}
\alias{study_design_arms_groups_and_interventions_fields}
\title{Study Design, Arms, Groups and Interventions Fields}
\format{
A character vector with a curated selection of 20 fields
}
\source{
\url https://prsinfo.clinicaltrials.gov/definitions.html#StudyDesign
}
\usage{
data(study_design_arms_groups_and_interventions_fields)
}
\description{
Character vector containing a selection of fields providing information about
the study design, number of arms or groups, and types of interventions involved.
}
\examples{
study_design_arms_groups_and_interventions_fields
}
\keyword{fields}
