% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_fields_vector_functions.R
\name{study_fields_vector}
\alias{study_fields_vector}
\title{Get All Study Fields Names}
\usage{
study_fields_vector()
}
\description{
This function returns a character vector with all 322 (May 2020) queriable field.
Uses glue.
names
}
\examples{
study_fields_vector()
}
\keyword{argument}
\keyword{fields}
