% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{for_netmeta}
\alias{for_netmeta}
\title{Data fields input for netmeta}
\format{
A character vector with a selection of 17 fields
}
\usage{
data(for_netmea)
}
\description{
Character vector containing .
}
\examples{
f
}
\keyword{fields}
