% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_fields.R
\name{get_study_fields}
\alias{get_study_fields}
\title{Get Over 20 Study Fields}
\usage{
get_study_fields(search_expr, fields, max_studies = 30)
}
\arguments{
\item{search_expr}{A string following the querying guidelines at
https://clinicaltrials.gov/api/gui/ref/syntax and https://clinicaltrials.gov/api/gui/ref/expr}

\item{fields}{A character vector with field names. All field names available at
https://clinicaltrials.gov/api/info/study_fields_list?fmt=JSON, with "all_fields",
or with "get_vector_all_study_fields()"}

\item{max_studies}{A number indicating how many studies of all the ones that
match the search expression will be returned.}
}
\description{
This function takes a search expression and character vector with any number
of field names and returns one dataframe with the merged and parsed csv API response.
}
\examples{
get_study_fields(expr = 'heart attack',fields = basic_info_and_results_fields())
}
\keyword{fields}
\keyword{get}
\keyword{many}
