% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors-help-data.R
\docType{data}
\name{identification_and_status_fields}
\alias{identification_and_status_fields}
\title{Study Identification and Status Fields}
\format{
A character vector
}
\source{
\url{https://prsinfo.clinicaltrials.gov/definitions.html#identification
}
\usage{
data(identification_and_status)
}
\description{
Character vector containing a small selection of fields including study identification, study status,
sponsor, regulatory oversight, relevant health condition and a brief summary and
detailed description.
}
\examples{
identification_and_status_fields
}
\keyword{fields}
