% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors_help-data.R
\docType{data}
\name{registration_fields}
\alias{registration_fields}
\title{Registration Data Elements Fields}
\format{
A character vector with a selection of 59 fields
}
\source{
\url
}
\usage{
data(registration_fields)
}
\description{
Character vector containing registration data elements fields, these elements are
all uploaded before the study is completed
}
\examples{
registration_fields
}
\keyword{fields}
