% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_vectors_help-data.R
\docType{data}
\name{results_fields}
\alias{results_fields}
\title{Results Data Elements Fields}
\format{
A character vector with a selection of 70 fields
}
\source{
\url
}
\usage{
data(registration_fields)
}
\description{
Character vector containing results data elements fields.
}
\examples{
results_fields
}
\keyword{fields}
