% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_vector_unique.R
\name{split_vector_unique}
\alias{split_vector_unique}
\title{Splits Fields Character Vector}
\usage{
split_vector_unique(x, n)
}
\arguments{
\item{x}{A character vector}

\item{n}{The number of strings in each column}
}
\description{
This function splits a character vector of any length into a matrix with n rows
of x elements. This is to get around the 20 field limit imposed by the API. Currently
duplicated strings are placed in new columns. FUnction could be simplified.
Created by user ngm Feb 2018 Stackoverflow
}
\examples{
split_vector_unique(x= sample(c("April", "is", "the","cruellest","month","breeding",
"lilacs"), size = 60, replace = TRUE), n =20)

}
\keyword{split}
\keyword{unique}
\keyword{vector}
